numeric variable types to standard names

+++ 01 Introductory Problems/01-weird-algorithm.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-    uint64_t n;
+    unsigned long long n;
     std::cin >> n;
 
     std::cout << n;
